Правки по замечаниям о делении.

+++ src/org/labs/lab1/Lab1.java
@@ -26,7 +26,7 @@ public class Lab1 {
 
     public static void divide() {
         int number = 10500;
-        int result = (number / 10) / 10;
+        double result = ((double)number / 10) / 10 ;
         System.out.println(result);
     }
